requirements.txt
streamlit
pillow
openai
